+++ content/graph/chapter.tex
@@ -40,9 +40,10 @@
 	\kactlimport{LCA.h}
 	\kactlimport{CompressTree.h}
 	\kactlimport{HLD.h}
-	\kactlimport{LinkCutTree.h}
 	\kactlimport{Kruskal.h}
 	\kactlimport{DirectedMST.h}
+	\kactlimport{LinkCutTree.h}
+	\kactlimport{LinkCutTree-PathUpdates.h}
 
 \section{Math}
 	\subsection{Number of Spanning Trees}
